backend/requirements.txt
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
psycopg[binary]==3.3.3
celery==5.6.3
redis==7.4.0
qdrant-client==1.17.1
requests==2.32.4
beautifulsoup4==4.13.4
readability-lxml==0.8.4.1
PyMuPDF==1.26.4
